requirements.txt
httpx==0.22.0
click==7.1.2
python-dotenv==0.5.1
numpy==1.23.5
click==8.0.0
python-dotenv==0.21.0
scikit-learn==0.24.1
dataclasses==0.6
pyyaml==6.0
marshmallow-dataclass==8.3.0
pandas==1.1.5
boto3==1.17.49
mlflow==1.15.0
mlflow==1.25.1
pytest==7.2.0
nbformat==5.1.3
hydra-core==1.2.0
omegaconf==2.2.3
opencv-python==4.6.0.66
protobuf==3.20.0
pandas_profiling==3.5.0
ipykernel==6.19.0
typing==3.7.4.3
wget==3.2
gdown==4.5.3
fastapi==0.85.1
fastapi-health==0.4.0
pydantic==1.10.2
uvicorn==0.19.0
uvicorn==0.19.0
httpx==0.22.0
botocore==1.20.112
requests==2.28.1
coverage==6.5.0
seaborn==0.12.1
